app/requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
databases[sqlite]==0.5.3
sqlalchemy==1.4.29
aiohttp==3.8.1
# only for testing purposes
pytest
requests
python-decouple